feat(ui): déplacer cas-usage list vers dossiers/[id] + gestion draft
